clean up a bunch of code, removing old print statements, fixing typos, reorganizing tests
