+++ ciderpress/gpaw/nldf_interface.py
@@ -540,43 +540,3 @@ def wrap_fft_mpi(pd):
     fft_obj.initialize_backend()
     wrapper = FFTWrapper(fft_obj, distribution, pd)
     return wrapper
-
-
-if __name__ == "__main__":
-    from gpaw.mpi import MPIComm, world
-
-    np.random.seed(98)
-
-    comm = MPIComm(world)
-    N_c = [100, 80, 110]
-    cell_cv = np.diag(np.ones(3))
-    ciderpw = LibCiderPW(N_c, cell_cv, comm)
-    ciderpw.initialize_backend()
-
-    from gpaw.pw.descriptor import PWDescriptor
-
-    gd = GridDescriptor(N_c, cell_cv, comm=MPIComm(world))
-    pd = PWDescriptor(180, gd, dtype=float)
-
-    wrapper = wrap_fft_mpi(pd)
-
-    input = np.random.normal(size=N_c)[None, :, :, :]
-    input = gd.distribute(input)
-
-    output_ref = pd.fft(input)
-    sum = np.abs(output_ref).sum()
-    sum = pd.gd.comm.sum_scalar(sum)
-    import time
-
-    t0 = time.monotonic()
-    for i in range(20):
-        output_test = wrapper.fft(input)
-    t1 = time.monotonic()
-    print("TIME", t1 - t0)
-    input_test = wrapper.ifft(output_test)
-    input_ref = pd.ifft(output_ref)
-
-    from numpy.testing import assert_allclose
-
-    assert_allclose(output_test[0], output_ref)
-    assert_allclose(input_test[0], input_ref)
