+++ ciderpress/pyscf/tests/test_uks_grad.py
@@ -15,6 +15,11 @@ SETTINGS = {
     "xmix": 0.25,
 }
 
+NULL_SETTINGS = {
+    "xkernel": "GGA_X_PBE",
+    "ckernel": "GGA_C_PBE",
+    "xmix": 0.00,
+}
 
 XC_SETTINGS = {
     "xkernel": None,
@@ -49,7 +54,7 @@ def build_ks_calc(mol, mlfunc, df=False, alt_settings=None):
 
 
 def setUpModule():
-    global mol, mlfuncs, mf1, mf2, mf3, mf4, mf5
+    global mol, mlfuncs, mf1, mf2, mf3, mf4, mf5, mf6, mf7
     mlfuncs = [
         "functionals/{}.yaml".format(fname)
         for fname in [
@@ -57,7 +62,6 @@ def setUpModule():
             "CIDER23X_SL_MGGA",
             "CIDER23X_NL_GGA",
             "CIDER23X_NL_MGGA_DTR",
-            # "VB7_MAPPED",
         ]
     ]
     mol = gto.Mole()
@@ -82,12 +86,21 @@ def setUpModule():
     mf4.kernel()
     mf5 = build_ks_calc(mol, mlfuncs[3], df=False).density_fit()
     mf5.kernel()
+    mf6 = build_ks_calc(mol, mlfuncs[3], df=True, alt_settings=NULL_SETTINGS)
+    mf6.kernel()
+    mf7 = dft.UKS(mol).density_fit()
+    mf7.xc = "PBE"
+    mf7.grids.level = 1
+    mf7.conv_tol = CONV_TOL
+    mf7.kernel()
     if (
         not mf1.converged
         or not mf2.converged
         or not mf3.converged
         or not mf4.converged
         or not mf5.converged
+        or not mf6.converged
+        or not mf7.converged
     ):
         raise RuntimeError(
             "{} {} {} {} {}".format(
@@ -96,14 +109,16 @@ def setUpModule():
                 mf3.converged,
                 mf4.converged,
                 mf5.converged,
+                mf6.converged,
+                mf7.converged,
             )
         )
 
 
 def tearDownModule():
-    global mol, mlfuncs, mf1, mf2, mf3, mf4, mf5
+    global mol, mlfuncs, mf1, mf2, mf3, mf4, mf5, mf6, mf7
     mol.stdout.close()
-    del mol, mlfuncs, mf1, mf2, mf3, mf4, mf5
+    del mol, mlfuncs, mf1, mf2, mf3, mf4, mf5, mf6, mf7
 
 
 class KnownValues(unittest.TestCase):
@@ -163,6 +178,10 @@ class KnownValues(unittest.TestCase):
 
         g2 = mf5.nuc_grad_method().set(grid_response=True).kernel()
         assert_almost_equal(g2, g, 9)
+
+        g3 = mf6.nuc_grad_method().set(grid_response=True).kernel()
+        g4 = mf7.nuc_grad_method().set(grid_response=True).kernel()
+        assert_almost_equal(g3, g4, 9)
 
     def _check_fd(self, functional):
         import os
